requirements.txt
absl-py==2.1.0
aiofiles==23.2.1
aiohappyeyeballs==2.5.0
opencv-python
aiohttp==3.11.13
aiosignal==1.3.2
aliyun-python-sdk-core==2.16.0
aliyun-python-sdk-kms==2.16.5
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
arrow==1.3.0
astunparse==1.6.3
async-timeout==5.0.1
audioread==3.0.1
click==8.1.8
contourpy==1.3.1
coverage==5.5
crcmod==1.7
cryptography==44.0.2
cycler==0.12.1
dashscope==1.22.2
deprecation==2.1.0
editdistance==0.8.1
einops==0.8.1
eyed3==0.9.7
fastapi==0.115.11
ffmpy==0.5.0
filelock==3.17.0
filetype==1.2.0
flatbuffers==25.2.10
fonttools==4.56.0
fqdn==1.5.1
frozenlist==1.5.0
fsspec==2025.2.0
funasr==1.2.6
gast==0.6.0
google-pasta==0.2.0
gradio==5.20.0
gradio_client==1.7.2
groovy==0.1.2
grpcio==1.71.0
h5py==3.13.0
huggingface==0.0.1
huggingface-hub==0.29.1
hydra-core==1.3.2
idna==3.10
intel-openmp==2021.4.0
ipywidgets==8.1.5
isoduration==20.11.0
jaconv==0.4.0
jamo==0.4.1
jieba==0.42.1
jmespath==0.10.0
joblib==1.4.2
json5==0.10.0
jsonpointer==3.0.0
jupyter==1.1.1
jupyter-console==6.6.3
jupyterlab_widgets==3.0.13
kaldiio==2.18.0
keras==3.9.0
kiwisolver==1.4.8
lazy_loader==0.4
libclang==18.1.1
librosa==0.10.2.post1
llvmlite==0.44.0
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.10.1
mdurl==0.1.2
mkl==2021.4.0
ml_dtypes==0.5.1
modelscope==1.23.1
more-itertools==10.6.0
mpmath==1.3.0
msgpack==1.1.0
multidict==6.1.0
namex==0.0.8
networkx==3.4.2
numba==0.61.0
numpy==1.26.4
omegaconf==2.3.0
openai-whisper==20240930
opt_einsum==3.4.0
optree==0.14.1
orjson==3.10.15
oss2==2.19.1
pandas==2.2.3
pillow==11.1.0
playsound==1.3.0
pooch==1.8.2
propcache==0.3.0
protobuf==5.29.3
pyAudioAnalysis==0.3.14
pycryptodome==3.21.0
pydantic==2.10.6
pydantic_core==2.27.2
pydub==0.25.1
pynndescent==0.5.13
pyparsing==3.2.1
python-multipart==0.0.20
pytorch-wpe==0.0.1
pytz==2025.1
regex==2024.11.6
rich==13.9.4
rotary-embedding-torch==0.8.6
ruff==0.9.9
safehttpx==0.1.6
scikit-learn==1.6.1
scipy==1.15.2
semantic-version==2.10.0
sentencepiece==0.2.0
shellingham==1.5.4
soundfile==0.13.1
soxr==0.5.0.post1
starlette==0.46.0
sympy==1.13.3
tbb==2021.13.1
tensorboard==2.19.0
tensorboard-data-server==0.7.2
tensorboardX==2.6.2.2
tensorflow==2.19.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.5.0
threadpoolctl==3.5.0
tiktoken==0.9.0
toml==0.10.2
tomlkit==0.13.2
torch==2.3.0
torch-complex==0.4.4
torchaudio==2.3.0
tqdm==4.67.1
typer==0.15.2
types-python-dateutil==2.9.0.20241206
tzdata==2025.1
umap-learn==0.5.7
uri-template==1.3.0
uvicorn==0.34.0
wcwidth==0.2.13
webcolors==24.11.1
webencodings==0.5.1
websockets==15.0
Werkzeug==3.1.3
widgetsnbextension==4.0.13
wrapt==1.17.2
yarl==1.18.3
